refactor: add app layer architecture skeleton
