+++ src/app/providers/AppProviders.tsx
@@ -1,0 +1,11 @@
+import type { PropsWithChildren } from 'react'
+import { AuthProvider } from '../../auth/AuthContext'
+import { ThemeProvider } from '../../theme/ThemeProvider'
+
+export function AppProviders({ children }: PropsWithChildren) {
+  return (
+    <ThemeProvider>
+      <AuthProvider>{children}</AuthProvider>
+    </ThemeProvider>
+  )
+}
