+++ src/app/providers/index.ts
@@ -1,0 +1,1 @@
+export { AppProviders } from './AppProviders'
